requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.34.3
requests>=2.32.3
fake-useragent>=2.2.0
sqlalchemy>=2.0.41
aiosqlite>=0.20.0
greenlet>=3.2.2

chardet==5.2.0
aiohttp==3.11.18
loguru==0.7.3
pydantic==2.11.4
beautifulsoup4==4.13.4
bs4==0.0.2

ruamel.yaml==0.18.14
openai==1.97.0
